+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file
 
+## 0.0.12-pre - 2026-05-12
+
+- add SpeedUnitAnnex compat
+
+
 ## 0.0.11-pre - 2026-05-08
 
 - use LZF compression for messages